backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-extensions==3.2.3
python-decouple==3.8
Pillow==10.1.0
psycopg2-binary==2.9.9
django-filter==23.3
django-crispy-forms==2.1
crispy-bootstrap5==0.7
celery==5.3.4
redis==5.0.1
django-redis==5.4.0
whitenoise==6.6.0
gunicorn==21.2.0

